Created Digit_Counter
Output

For Positive Number
Input: 1234
Output: The number of digits in 1234 is: 4
For 0
Input: 0
Output: The number of digits in 0 is: 1
For Negative Number
Input: -12345
Output: The number of digits in -12345 is: 5
Corner Cases You Might Miss: We have used the loop and carried on iterations till the number becomes 0. What if the number was already 0? It still has 1 digit. So, we have handled that separately. Also, to avoid any confusion, the negative numbers are converted to positive in our function and then we calculate their number of digits.
Time Complexity: O(log10N) where N is the input number. This is because we keep dividing the number by 10.
Auxiliary Space: O(1) as we have not used any extra space.