Created Number Finder and Counter
Corner Cases You Might Miss: If the input number is 0 and the digit is also 0, it becomes a crucial corner case. This is because the number 0 has 1 frequency of digit 0 but it will not be handled correctly by our loop. So, we do this separately. Also, we have converted the negative numbers to positive ones to solve this problem.
Time Complexity: O(log10N) where N is the input number. This is because we keep dividing the number by 10.
Auxiliary Space: We have not used any auxiliary Space here. So, it is O(1).