Created Mystery Number
A number N is said to be a mystery number if the number can be expressed by the sum of two numbers. Note that these two numbers must be reverse of each other. It lies between 22 to 198, i.e. 22<=N<=198.

It is called mystery numbers because these are the multiples of 11, and the sum of place values of unit (1) and tens (10) place is 11 in decimal numeration.